+++ lib/Models/Terria.ts
@@ -37,6 +37,7 @@ import TimeVarying from "../ModelMixins/TimeVarying";
 import { BaseMapViewModel } from "../ViewModels/BaseMapViewModel";
 import TerriaViewer from "../ViewModels/TerriaViewer";
 import { HelpContentItem } from "../ReactViewModels/defaultHelpContent";
+import { defaultTerms, Term } from "../ReactViewModels/defaultTerms";
 import CameraView from "./CameraView";
 import CatalogGroup from "./CatalogGroupNew";
 import CatalogMemberFactory from "./CatalogMemberFactory";
@@ -93,6 +94,7 @@ interface ConfigParameters {
   showWelcomeMessage?: boolean;
   showInAppGuides?: boolean;
   helpContent?: HelpContentItem[];
+  helpContentTerms?: Term[];
 }
 
 interface StartOptions {
@@ -205,7 +207,8 @@ export default class Terria {
     globalDisclaimer: undefined,
     showWelcomeMessage: false,
     showInAppGuides: false,
-    helpContent: []
+    helpContent: [],
+    helpContentTerms: defaultTerms
   };
 
   @observable